### 2025-05-05 #703 @marcboschmatas
partialling.out
Submitting Author Name: Marc Bosch Matas
Submitting Author Github Handle: @mboschmatas
Repository:  https://github.com/marcboschmatas/partialling.out
Version submitted: 0.1.1
Submission type: Stats
Badge grade: bronze

Version accepted: TBD
---

```
Package: partialling.out
Title: Residuals from partial regressions
Version: 0.0.1
    person("Marc", "Bosch-Matas", , "mboschmatas@gmail.com", role = c("aut", "cre"),
Description: Creates a data frame with the residuals of partial regressions of the main explanatory variable and the variable of interest. This method follows the Frisch-Waugh-Lovell theorem, as explained in Lovell (2008) <doi:10.3200/JECE.39.1.88-91>.
    tinytest,
    palmerpenguins,
    tinytable,
    tinyplot,
    fwlplot
Roxygen:list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    glue,
    lifecycle,
    fixest,
    lfe,
    stats
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [ ] A pre-submission inquiry has been approved in [issue#<issue_num>](#695 )

Econometricians who intend to visualise models via scatter plots or similar tools. Transform multivariate into bivariate models for simpler visualisation or analysis.

    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

[fwlplot](https://github.com/kylebutts/fwlplot/tree/main/R)

Bronze
