### 2025-05-20 #703 @mpadge
@marcboschmatas We're please to announce that @nhejazi has just joined us as our new statistical software editor, and will be handling this as his very first submission. (So please be especially kind and considerate 😄 )

@nhejazi Our general procedures are described [in the editors chapter of our _Dev Guide_](https://devguide.ropensci.org/editortemplate.html). You main tasks from here are:

1. Fill our and paste the [editors template](https://devguide.ropensci.org/editortemplate.html), which will also provide an opportunity for you to familiarize yourself with the package. Examples of typical comments can be seen throughout the [list of issues](https://github.com/ropensci/software-review/issues?q=sort%3Aupdated-desc%20is%3Aissue%20%22Editor%20comments%22). All stats packages first pass through a pre-submission inquiry, which is this case was #695. A quick look over that will likely be helpful too.
2. Find and assign two reviewers, as described in [this _Dev Guide_ section](https://devguide.ropensci.org/softwarereview_editor.html#look-for-and-assign-two-reviewers).

Feel free to ask any questions you may have here, or reach out privately if you prefer. Welcome aboard Nima, and thanks for the exciting submission Marc!
