### 2025-06-03 #703 @nhejazi
for now, i'm pasting the editor checklist as noted above and will return to it as the package review process gets underway

---

---
