+++ README.md
@@ -12,6 +12,8 @@ status](https://www.r-pkg.org/badges/version/partialling.out)](https://CRAN.R-pr
 [![Codecov test
 coverage](https://codecov.io/gh/marcboschmatas/partialling.out/graph/badge.svg)](https://app.codecov.io/gh/marcboschmatas/partialling.out)
 [![R-CMD-check](https://github.com/marcboschmatas/partialling.out/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/marcboschmatas/partialling.out/actions/workflows/R-CMD-check.yaml)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/703_status.svg)](https://github.com/ropensci/software-review/issues/703)
 <!-- badges: end -->
 
 Partialling out is a package that allows to generate residualised
@@ -23,12 +25,12 @@ already been implemented in
 [`fwlplot`](https://github.com/kylebutts/fwlplot), this package offers
 three new characteristics.
 
-- It uses an already existing model instead of a formula.
+  - It uses an already existing model instead of a formula.
 
-- Works with `lm` and `felm` objects alongside `feols`.
+  - Works with `lm` and `felm` objects alongside `feols`.
 
-- Returns a data.frame with residualised variables instead of a plot,
-  thus offering more freedom of what to do with the results.
+  - Returns a data.frame with residualised variables instead of a plot,
+    thus offering more freedom of what to do with the results.
 
 ## Installation
 
@@ -51,7 +53,7 @@ library(tinytable)
 library(tinyplot)
 library(palmerpenguins)
 
-model <- lm(bill_length_mm ~  bill_depth_mm + species, data = penguins)
+model <- lm(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 summary(model)
 #> 
 #> Call:
@@ -78,7 +80,7 @@ summary(model)
 
 Using the `partialling_out` function, you can get the residualised
 variable of interest (bill length) and of the first explanatory variable
-(bill_length), i.e. it would return the residuals of the following two
+(bill\_length), i.e. it would return the residuals of the following two
 regressions.
 
 ``` r
@@ -96,32 +98,50 @@ tt(head(res)) |>
   style_tt(align = "c")
 ```
 
-| res_bill_length_mm | res_bill_depth_mm |
-|--------------------|-------------------|
-| 0.31               | 0.35              |
-| 0.71               | -0.95             |
-| 1.51               | -0.35             |
-| -2.09              | 0.95              |
-| 0.51               | 2.25              |
-| 0.11               | -0.55             |
+| res\_bill\_length\_mm | res\_bill\_depth\_mm |
+| --------------------- | -------------------- |
+| 0.31                  | 0.35                 |
+| 0.71                  | \-0.95               |
+| 1.51                  | \-0.35               |
+| \-2.09                | 0.95                 |
+| 0.51                  | 2.25                 |
+| 0.11                  | \-0.55               |
 
 ## Checking the results
 
[…]
