+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   partialling_out: partialling_out.html
-last_built: 2025-04-18T19:12Z
+last_built: 2025-06-06T06:51Z
 urls:
   reference: https://marcboschmatas.github.io/partialling.out/reference
   article: https://marcboschmatas.github.io/partialling.out/articles
