### 2025-07-07 #703 @ctesta01
Estimated hours spent reviewing: ~3

---

Checklist from the **General Guidelines** at <https://devguide.ropensci.org/softwarereview_reviewer.html>:

  - Does the code comply with general principles in the Mozilla reviewing guide?
    * Yes
  - Does the package comply with the rOpenSci packaging guide?
    * Yes
  - Are there improvements that could be made to the code style?
    * None necessary.
    * Minor/trivial suggestions: There are a few places where `}else` appears and is lacking a space between the closing brace and `else` (as in I suggest it should be `} else`).  The latter is displayed as the standard in the `tidyverse` style guide, though without the space is not explicitly listed as "Bad" either.  My perspective is that it is more common/standard to have a space between them.
    * The `tidyverse` style guide says to "Use vertical whitespace sparingly" and "Only use a single empty line when needed to separate functions or pipes." There are several places that many empty lines are inserted, which could be removed to follow this principle.
  - Is there code duplication in the package that should be reduced?
    * No
  - Are there user interface improvements that could be made?
    * No
  - Are there performance improvements that could be made?
    * No
  - Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient? Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?
    * Yes.  Whether users encounter the README, vignettes, or function documentation first, the documentation is clear and sufficient.
  - Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?
    * Yes, the S3 OOP system is used effectively so that users only need to call the `partialling_out` S3 dispatch method provided by the package.
  - If you have your own relevant data/problem, work through it with the package. You may find rough edges and use-cases the author didn’t think about.
    * Suggestion: I have often used polynomials or the "AsIs" operator `I()` regression formulas. I tried the following and was met with the following errors:
```
# using I() in an lm() results in an unexpected error
> model <- lm(bill_length_mm ~ bill_depth_mm + species + I(flipper_length_mm^2), data = penguins)
> res <- partialling_out(model, data = penguins)
Error in `data[, formulas$filter_terms]`:
! Can't subset columns that don't exist.
✖ Column `I(flipper_length_mm^2)` doesn't exist.

# using poly() in lm() results in an error
> model <- lm(hp ~ poly(cyl, 2) + disp, data = mtcars)
> res <- partialling_out(model, data = mtcars)
Error in `[.data.frame`(data, , formulas$filter_terms) :
  undefined columns selected
```
You may want to provide similar errors as to your error documented in the test "expect error if interaction terms with i() are passed in fixest" for these scenarios. It may be worth checking the documentation for `?formula` as there are some fairly obscure features formulas parsed by `lm()` support including `%in%` or `/` to make sure those are functioning correctly.

It may be worth pointing out to users that they need to have `pak` installed and GitHub credentials set up through `{gitcreds}` and/or `usethis::create_github_token()`.  I had these things set up, so installation worked fine for me, but if a user doesn't have GitHub credentials already registered with R, they will face some errors using `{pak}` to install the package.

#### Contributing

I noticed the contribution guidelines are in `.github/contributing.md` -- as a suggestion, it would be reasonable to point this out in the README as it is not immediately obvious where they are.

#### Math Display Issue

While the mathematical equations render nicely in the README when viewed on GitHub, they do not appear rendered correctly on the documentation website <https://marcboschmatas.github.io/partialling.out/index.html>.

I'm not exactly certain how this could be solved, as the use of `$$` to display equations seems reasonable in the README.Rmd file.

I have run into a similar issue before with math rendering poorly in `{pkgdown}` generated pages (though in my situation it was from Roxygen generated documentation) and following the advice in this thread <https://github.com/r-lib/pkgdown/issues/2704> helped me (specifically louisaslett's comment on Aug 23 2024 <https://github.com/r-lib/pkgdown/issues/2704#issuecomment-2307055568>).

#### Overall Comment

This looks like a useful contribution for users interested in residualizing variables to perform Frisch Waugh Lovell style calculations with model objects and fixed effects models.  Kudos on the work and your package submission.
