### 2025-07-13 24acf53
added explicit errors for AsIs and poly and modified test suite
