+++ inst/tinytest/test_partialling.out.R
@@ -1,4 +1,3 @@
-
 # Placeholder with simple test
 library(palmerpenguins)
 library(fixest)
@@ -22,7 +21,6 @@ set.seed(1234)
 NULL
 # nolint end
 
-
 # Test for errors warnings etc ----
 
 # nolint start
@@ -33,161 +31,172 @@ NULL
 
 # nolint end
 
-
 ## Expect error if no data is provided ----
 
 ### lm ----
 safe_partialling_out <- safely(partialling_out)
 
 
-mod <- lm(bill_length_mm ~  bill_depth_mm + species,
-  data = penguins)
+mod <- lm(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 
 
 lmerrornodata <- safe_partialling_out(mod)
 
-expect_equal(lmerrornodata$error$message,
-  "No data has been provided")
-
+expect_equal(lmerrornodata$error$message, "No data has been provided")
 
 
 ### feols ----
 
-mod <- feols(bill_length_mm ~  bill_depth_mm | species,
-  data = penguins)
+mod <- feols(bill_length_mm ~ bill_depth_mm | species, data = penguins)
 
 feolserrornodata <- safe_partialling_out(mod)
 
-expect_equal(feolserrornodata$error$message,
-    "No data has been provided")
+expect_equal(feolserrornodata$error$message, "No data has been provided")
 
 
 ### felm----
-mod <- felm(bill_length_mm ~  bill_depth_mm | species,
-  data = penguins)
+mod <- felm(bill_length_mm ~ bill_depth_mm | species, data = penguins)
 
 felmerrornodata <- safe_partialling_out(mod)
 
-expect_equal(felmerrornodata$error$message,
-    "No data has been provided")
+expect_equal(felmerrornodata$error$message, "No data has been provided")
 
 ## expect error if data is not a data.frame ----
 
 ### lm ----
-mod <- lm(bill_length_mm ~  bill_depth_mm + species,
-  data = penguins)
+mod <- lm(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 
 lmerrornodf <- safe_partialling_out(mod, data = c(1, 2, 3))
 
-expect_equal(lmerrornodf$error$message,
-  "data input should be a data.frame")
+expect_equal(lmerrornodf$error$message, "data input should be a data.frame")
 
 
 ### feols ----
 
-mod <- feols(bill_length_mm ~  bill_depth_mm + species,
-  data = penguins)
+mod <- feols(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 
 feolserrornodf <- safe_partialling_out(mod, data = c(1, 2, 3))
 
-expect_equal(feolserrornodf$error$message,
-  "data input should be a data.frame")
+expect_equal(feolserrornodf$error$message, "data input should be a data.frame")
 
 
 ### felm ----
 
-mod <- felm(bill_length_mm ~  bill_depth_mm + species,
-  data = penguins)
+mod <- felm(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 
 felmerrornodf <- safe_partialling_out(mod, data = c(1, 2, 3))
 
-expect_equal(felmerrornodf$error$message,
-  "data input should be a data.frame")
+expect_equal(felmerrornodf$error$message, "data input should be a data.frame")
 
 
 ## expect error if weights are non numeric ----
 
 ### lm ----
-mod <- lm(bill_length_mm ~  bill_depth_mm + species, data = penguins)
+mod <- lm(bill_length_mm ~ bill_depth_mm + species, data = penguins)
 
-nonnumweightslm <- safe_partialling_out(mod, data = penguins, weights = rep("a", nrow(penguins)))
[…]
