+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   partialling_out: partialling_out.html
-last_built: 2025-07-13T17:23Z
+last_built: 2025-07-13T17:31Z
 urls:
   reference: https://marcboschmatas.github.io/partialling.out/reference
   article: https://marcboschmatas.github.io/partialling.out/articles
