### 2025-07-14 4d694a8
pkgdown math rendering
