### 2025-07-14 #703 @marcboschmatas
I've gone through most of the stuff you're commenting here

>         Minor/trivial suggestions: There are a few places where }else appears and is lacking a space between the closing brace and else (as in I suggest it should be } else). The latter is displayed as the standard in the tidyverse style guide, though without the space is not explicitly listed as "Bad" either. My perspective is that it is more common/standard to have a space between them.
>         The tidyverse style guide says to "Use vertical whitespace sparingly" and "Only use a single empty line when needed to separate functions or pipes." There are several places that many empty lines are inserted, which could be removed to follow this principle.

Done.

> Suggestion: I have often used polynomials or the "AsIs" operator I() regression formulas. I tried the following and was met with the following errors:
>
>
> You may want to provide similar errors as to your error documented in the test "expect error if interaction terms with i() are passed in fixest" for these scenarios. It may be worth checking the documentation for ?formula as there are some fairly obscure features formulas parsed by lm() support including %in% or / to make sure those are functioning correctly.

I added explicit errors for the use of AsIs and polynomials, and tested them. They seem to throw the appropriate error

> Installation

> It may be worth pointing out to users that they need to have pak installed and GitHub credentials set up through {gitcreds} and/or usethis::create_github_token(). I had these things set up, so installation worked fine for me, but if a user doesn't have GitHub credentials already registered with R, they will face some errors using {pak} to install the package.

Need to check these problems with someone else's computer.

> Contributing
>
> I noticed the contribution guidelines are in .github/contributing.md -- as a suggestion, it would be reasonable to point this out in the README as it is not immediately obvious where they are.

Done

> Math Display Issue
I'm at a nearly total loss here, and any help will be very welcome. Apparently, the problem comes from the rendering of README.Rmd to README.md

An equation written as

`$$ X_2 = X_1 \beta_1 + u $$`

gets turned into this, in README.md

```

![ X\_2 = X\_1 \\beta\_1 + u
](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;%20X_2%20%3D%20X_1%20%5Cbeta_1%20%2B%20u%20
" X_2 = X_1 \\beta_1 + u ")
```

Any suggestions forward are very welcome. Perhaps I'll have to open an issue in devtools to see what's going on.
