### 2025-07-14 #703 @marcboschmatas
I'll give that a go, thanks a lot!
Also, thanks for the review and the offer. I'll have to look very closely on your formula parsing. I know that my `prepare_formula()` will need a refactor sooner or later, so any suggestions welcome.
