### 2025-07-14 #703 @nhejazi
hi @aloy -- thank you for offering to review the `partialling.out` package submitted to rOpenSci.

the goals and scope of rOpenSci's review process are documented at [stats-devguide.ropensci.org/overview.html#overview-scope](https://stats-devguide.ropensci.org/overview.html#overview-scope).
