### 2025-07-15 #703 @ctesta01
Hi @marcboschmatas,

It looks like you've addressed everything I mentioned in my review well and have a plan for the remaining two pieces. On those two pieces that you're still working on, I have some comments:

## `pak` and GitHub PATs

I'm not exactly sure myself when `pak::pak` requires a GitHub PAT -- I'm thinking now this was something specific to my machine.  I just tried to install `{partialling.out}` on another machine with `pak` and didn't run into any trouble even though that machine doesn't have a GitHub PAT setup.  (Details:  I suspect what happened on my machine is that my PAT had expired and therefore I had to replace it;  most users won't face this type of an issue if they either don't have a GitHub PAT set up or have one that is currently valid at the time of install).  I think it would be fine to leave it as `pak::pak("marcboschmatas/partialling.out")` or if you're concerned about this, you could of course use `remotes::install_github` instead.

## math display issues

Regarding the math rendering issue, I've tried to think of some options you might consider.  **I think the easiest might be to just change the `$$ $$` display equations to `$ $` inline equations** (if that's okay with you / suits your taste) since the display equations in your readme aren't particularly long and it seems the inline equations work fine both on GitHub and the `pkgdown` website.

Some other options that I'm less optimistic about are:

  * to see if it's any better using Quarto and rendering a `format: gfm` document instead of using R Markdown — I'm not particularly convinced this would yield any different results, though, and haven't tried it myself.
  * I haven't figured this out for myself, but is there any way to get R Markdown or Quarto to render to a .md extension but leave the `$$ $$` and `$ $` equations as they are?  (i.e., don't convert them to images rendered by codecogs.)  My reasoning for suggesting this is that [GitHub has supported rendering math natively](https://github.blog/news-insights/product-news/math-support-in-markdown/) for a while (since May 2022) in markdown documents.  I think if the math wasn't rendered to images via codecogs, I think `pkgdown` would also handle the `$ $` and `$$ $$` math fine otherwise either via mathml, Mathjax or Katex (configurable according to instructions [here](https://pkgdown.r-lib.org/articles/customise.html#math-rendering)).

---

Glad to see new reviewers have joined on, and happy to see the progress on this!
