### 2025-07-16 #703 @marcboschmatas
Thanks @ctesta01 !
Interesting what you mention about `pak`. I asked some colleagues to try it and they had no problem.

Inline equations also throw out weird results. Right now, I've managed to have something a bit less of an eyesore, but I really need to keep trying. The idea of forcing README.md to keep `$$ $$` sounds like the easiest one. I tried `format: gfm` (still as a Rmd file, not quarto) and didn't work. Otherwise, I wonder if I can force `pkgdown` to get the homepage from README.Rmd instead of .md, which ought to work fine. Otherwise, jut revert to `pkgdown` 2.0.9 which used to work fine.
