### 2025-07-17 e4a9558
Fix README math
