+++ README.Rmd
@@ -61,11 +61,8 @@ summary(model)
 Using the `partialling_out` function, you can get the residualised variable of interest (bill length) and of the first explanatory variable (bill_length), i.e. it would return the residuals of the following two regressions.
 
 ```{r eval = FALSE}
-
 modely <- lm(bill_length_mm ~ species, data = penguins)
-
 modelx <- lm(bill_depth_mm ~ species, data = penguins)
-
 ```
 
 
@@ -83,29 +80,30 @@ tt(head(res)) |>
 
 The Frisch-Waugh-Lovell theorem states that for a linear model
 
-$$
+``` math
 Y = X_0 \beta_1 + X_2\beta_2 + u
-$$
+```
 
 The coefficient $\beta_2$ will be equivalent to that in the regression
 
-$$
+``` math
 M_{X_1} Y = M_{X_1}X_2\beta_2 + M_{X_1}u
-$$
+```
 
 Where $M_{X_1}Y$ are the residuals of the model
 
-$$
+``` math
 Y = X_1 \beta_1 + u
-$$
+```
 
 And $M_{X_1}X_2$ those of 
 
 
-$$
+``` math
 X_2 = X_1 \beta_1 + u
-$$
+```
 
+<br/>
 Accordingly, the coefficient of `res_bill_depth_mm` in the model `lm(res_bill_length_mm ~ res_bill_depth_mm)` will be the same of the coefficient of `bill_depth_mm` in the original model.
 
 ```{r}
