+++ README.md
@@ -25,12 +25,12 @@ already been implemented in
 [`fwlplot`](https://github.com/kylebutts/fwlplot), this package offers
 three new characteristics.
 
-  - It uses an already existing model instead of a formula.
+- It uses an already existing model instead of a formula.
 
-  - Works with `lm` and `felm` objects alongside `feols`.
+- Works with `lm` and `felm` objects alongside `feols`.
 
-  - Returns a data.frame with residualised variables instead of a plot,
-    thus offering more freedom of what to do with the results.
+- Returns a data.frame with residualised variables instead of a plot,
+  thus offering more freedom of what to do with the results.
 
 ## Installation
 
@@ -80,7 +80,7 @@ summary(model)
 
 Using the `partialling_out` function, you can get the residualised
 variable of interest (bill length) and of the first explanatory variable
-(bill\_length), i.e. it would return the residuals of the following two
+(bill_length), i.e. it would return the residuals of the following two
 regressions.
 
 ``` r
@@ -98,59 +98,42 @@ tt(head(res)) |>
   style_tt(align = "c")
 ```
 
-| res\_bill\_length\_mm | res\_bill\_depth\_mm |
-| --------------------- | -------------------- |
-| 0.31                  | 0.35                 |
-| 0.71                  | \-0.95               |
-| 1.51                  | \-0.35               |
-| \-2.09                | 0.95                 |
-| 0.51                  | 2.25                 |
-| 0.11                  | \-0.55               |
+| res_bill_length_mm | res_bill_depth_mm |
+|--------------------|-------------------|
+| 0.31               | 0.35              |
+| 0.71               | -0.95             |
+| 1.51               | -0.35             |
+| -2.09              | 0.95              |
+| 0.51               | 2.25              |
+| 0.11               | -0.55             |
 
 ## Checking the results
 
 The Frisch-Waugh-Lovell theorem states that for a linear model
 
-  
-![&#10;Y = X\_0 \\beta\_1 + X\_2\\beta\_2 +
-u&#10;](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;%0AY%20%3D%20X_0%20%5Cbeta_1%20%2B%20X_2%5Cbeta_2%20%2B%20u%0A
-"
-Y = X_0 \\beta_1 + X_2\\beta_2 + u
-")  
+``` math
+Y = X_0 \beta_1 + X_2\beta_2 + u
+```
 
-The coefficient
-![\\beta\_2](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;%5Cbeta_2
-"\\beta_2") will be equivalent to that in the regression
+The coefficient $\beta_2$ will be equivalent to that in the regression
 
-  
-![&#10;M\_{X\_1} Y = M\_{X\_1}X\_2\\beta\_2 +
-M\_{X\_1}u&#10;](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;%0AM_%7BX_1%7D%20Y%20%3D%20M_%7BX_1%7DX_2%5Cbeta_2%20%2B%20M_%7BX_1%7Du%0A
-"
-M_{X_1} Y = M_{X_1}X_2\\beta_2 + M_{X_1}u
-")  
+``` math
[…]
