### 2025-07-17 #703 @marcboschmatas
**Readme**
I approved the PR right now, this sounds perfect

**Computation efficiency**
Thanks a lot for your comments. I need to have a good look at it and see how it applies. I hope you don't mind if I ask you a few questions during the process. I'm OK with depending on fixest so I'll take a look.

**Plot methods**
I guess this falls into the nice-to-have category, but it wouldn't be too hard I guess. I'll have a good look at it during the weekend.
