+++ NAMESPACE
@@ -10,6 +10,8 @@ importFrom(glue,glue)
 importFrom(lfe,felm)
 importFrom(lifecycle,deprecated)
 importFrom(stats,as.formula)
+importFrom(stats,coef)
 importFrom(stats,formula)
 importFrom(stats,lm)
+importFrom(stats,resid)
 importFrom(stats,terms)
