### 2025-08-21 84e6332
plotting method testing & doc
