+++ inst/tinytest/helpers.R
@@ -1,0 +1,11 @@
+library(tinytest)
+library(tinysnapshot)
+
+
+options("tinysnapshot_os" = "Linux")
+options("tinysnapshot_device" = "svglite")
+options(
+  "tinysnapshot_device_args" = list(
+    user_fonts = fontquiver::font_families("Liberation")
+  )
+)
