### 2025-08-21 #703 @marcboschmatas
That's right. At the moment I've left it as `plot_partial_residuals()` (and renamed the class to `partial_residuals`), and added several options for plotting with a tinyplot backend. Also added snapshot tests and an explanation of the plotting function in the vignettes, so we have that covered, I guess.
