### 2025-09-03 #703 @nhejazi
@ctesta01, great--thank you for confirming

@kylebutts, thank you for adding the review checklist. it looks like there are two items unchecked--one on COIs, which i'd assume should be checked, and the other on community guidelines listed for contributing to the package. the latter may be addressed by the document linked to in the contributing section of the readme, but whether that is satisfactory is for you to assess
