### 2025-09-22 #703 @nhejazi
ok, thanks @marcboschmatas. it looks like there are some unchecked items in @aloy's checklist but i agree that we can move ahead without those items if we don't hear back soon.
