### 2025-10-15 #703 @marcboschmatas
Hi,
I still haven't got an invitation to slack, I guess it's coming soon. I'll start working on the post and PR once I have it ready!
